Cleaning up code and adding README, Dockerfile, and sample output
In this commit, I added metadata about some things that I thought would be important to the project. I also was able to clean up a lot of the unused code and removed a lot of the unecessary print statements (debugging purposes). This commit will also include an in depth readme of the project and the issues faced and a Dockerfile.